Refactored parsing of ZXT_A01 and fixed failing tests
